backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
asyncpg==0.29.0
pgvector==0.2.4
alembic==1.13.1

# Document Processing
pikepdf==8.10.1
pymupdf==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
pillow==10.2.0
PyPDF2==3.0.1
pdfplumber==0.10.3

# Google Gemini AI
google-generativeai==0.3.2

# Local embeddings - FIXED
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
sentence-transformers==2.3.1
transformers==4.36.2

# Other AI services
tiktoken==0.5.2
cohere==4.43

# Caching
redis==5.0.1

# HTTP
httpx==0.27.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
aiosmtplib==3.0.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3
loguru==0.7.2
markdown==3.5.1
email-validator==2.2.0
filetype==1.2.0
numpy==1.26.3

# Development/Testing
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.11
mypy==1.8.0
reportlab==4.0.7
psycopg2-binary>=2.9.9
